rename(camelCase) files, move to lib folder, and use pg.extra for tests
