+++ src/lib/trimIndent.js
